Add getPurchaseCancellationFlowType to lib/purchases/utils.ts
